new zip folder structure

+++ node_scripts/mkzip.js
@@ -19,7 +19,8 @@ var nwpath={
 	'linux':'linux-ia32'	
 }
 
-var platform=process.platform;
+
+
 var ZipWriter = require("./zipwriter").ZipWriter;
 var zip = new ZipWriter();
 var walk = function(dir) {
@@ -40,21 +41,25 @@ var walk = function(dir) {
     return results
 }
 var addfile=function(f,upfolder) {
+	
 	if (f.indexOf(".git")>-1 || f.indexOf(".bak")>-1  || f.indexOf(".log")>-1) {
 //		console.log('skip',f);
 		return;
 	}
-	console.log('add ',f);	
+	
 	var target=f;
-	if (upfolder) {
-		var folders=f.split('/');
+	var folders=f.split('/');
+	if (typeof upfolder=="number") {
 		while (upfolder--) folders.shift();
 		if (folders.length<1) {
 			throw "reduce the upfolder number"
 		}
 		target=folders.join('/');
+	} else if (typeof upfolder=="string") { //is targetfolder
+		target=upfolder+'/'+folders[folders.length-1];
 	}
 	zip.addFile(target,f);
+	console.log('add ',f, 'to',target);
 }
 var addtozip=function(files,upfolder) {
 	for (var i in files) {
@@ -78,26 +83,42 @@ var indexhtml='<html>\n<head>\n<meta charset="utf-8" />\n'+
 						'<link type="text/css" rel="stylesheet" href="build.css"/></head>\n'+
 						'<div id="main"></div>\n'+
 						'<script>window.nodeRequire=require;</script>\n'+
+						'<script src="jquery.js"></script>\n'+
+						'<script src="socket.io.js"></script>\n'+
+						'<script src="react-with-addons.js"></script>\n'+
 						'<script src="build.js"></script>\n'+
 						'</html>';
 var add_appfiles=function(appfolder,zip,extras) {
-	zip.addData("index.html",indexhtml);
-	addfile(appfolder+"/index.css",1);
-	addfile(appfolder+"/package.json",1);
-	addfile(appfolder+"/build/build.js",2);
-	addfile(appfolder+"/build/build.css",2);
+	zip.addData(appfolder+"/index.html",indexhtml);
+	addfile(appfolder+"/index.css");
+	
+	addfile(appfolder+"/build/build.js",appfolder);
+	addfile(appfolder+"/build/build.css",appfolder);
+	
 	if (extras) {
-		if (extras.files) addtozip(extras.files , 1);
+		if (extras.repos) {
+			console.log("add repos",extras.repos);
+			addtozip(extras.repos);
+		}
+		if (extras.files) {
+			for (var i=0;i<extras.files.length;i++) {
+				var f=extras.files[i];
+				if (typeof f=="string") {
+					addfile(f,appfolder); //default app folder
+				} else {
+					addfile(f[0],f[1]);//user specified folder
+				}
+			}
+		}
 		//array of node modules 
 		//['node_modules/yadb','node_modules/yase']
-		if (extras.repos) addtozip(extras.repos);
+		
 	}
 }
 
 var add_node_webkit=function() {
-	console.log('node_webkit/'+nwpath[platform],process.cwd())
-
-	addtozip(['node_webkit/'+nwpath[platform]],2);
+	//console.log('node_webkit/'+nwpath[platform],process.cwd()
+	for (var i in platform) addtozip(['node_webkit/'+nwpath[platform[i]]],1);
 	
 }
 /*
@@ -117,38 +138,49 @@ var changeSettings=function(fn) {
 	} else return null;
 }
 
-var changePackageJson=function(fn) {
+var changePackageJson=function(fn,appfolder) {
 	if (!fs.existsSync(fn)) return null;
 	var packagejson=JSON.parse(fs.readFileSync(fn,'utf8'));
-	var fullscreen=packagejson.window.toolbar;
-	if (fullscreen) {
-		packagejson.window.toolbar=false;
-		saveJson(fn,packagejson);
-		packagejson.window.toolbar=true;
-		return packagejson;
-	} else return null;
+	
+
+	var toolbar=packagejson.window.toolbar;
+	packagejson.window.toolbar=false;
+	packagejson.main="../"+appfolder+"/index.html";
+	saveJson(fn,packagejson);
+	packagejson.main="index.html";
+	packagejson.window.toolbar=true;
+
+	return packagejson;
 }
+
 var saveJson=function(fn,json) {
 	if (!fn || !json) return;
-	fs.writeFileSync(fn, JSON.stringify(json,'',' '),'utf8');
+	if (fs.existsSync(fn)) fs.writeFileSync(fn, JSON.stringify(json,'',' '),'utf8');
 }
+var platform=process.platform;
 var mkzip=function(appfolder,pf,product) {
+	console.log("appfolder",appfolder);
 	var starttime=new Date();
-	platform=pf || process.platform;
+	platform=pf || process.platform;		
+	if (platform.indexOf(",")>-1) platform=platform.split(",");
+	if (typeof platform=="string") platform=[platform];
 
 	var date =new Date().toISOString().substring(0,10);
 	var folders=appfolder.split(path.sep);
 	var appname=folders[folders.length-1];
 	var extras=null;
+	
 	if (fs.existsSync(appfolder+"/mkzip.json")){
 		extras=require(appfolder+"/mkzip.json");
 		product=product||extras.name||appname ;
 	} else {
 		product=product||appname;	
 	}
+	console.log(extras,appfolder+"/mkzip.json");
 	
 	var settings=changeSettings(appfolder+"/settings.json");
-	var packagejson=changePackageJson(appfolder+"/package.json");
+	var oripackagejson=changePackageJson(appfolder+"/package.json",appname);
+	for (var i in platform) addfile(appfolder+"/package.json",nwpath[platform[i]]);
 
 //	var shellscriptname='start-'+appname + shellscript[platform];
 	//create 
@@ -157,7 +189,7 @@ var mkzip=function(appfolder,pf,product) {
 	add_node_webkit();
 
 	console.log("");
-	var zipname=product +'-'+platform+'-'+date+'.zip';
+	var zipname=product +'-'+date+'.zip';
 	console.log('.....Creating Zip file.....'+zipname);
 	zip.saveAs(zipname,function() {
 	   console.log('time elapsed in seconds', Math.round(new Date()-starttime)/1000);
@@ -165,10 +197,8 @@ var mkzip=function(appfolder,pf,product) {
 	   console.log(zipname);
 	   process.chdir(appfolder);
 	   saveJson(appfolder+"/settings.json",settings);
-	   saveJson(appfolder+"/package.json",packagejson);
+	   saveJson(appfolder+"/package.json",oripackagejson);
 	});
-	
-
 }
